fix: 리펙토링 (#2)
